- update for 1.4 - split out the modules into separate files - remove wiki view show replacement and use hooks with javascript to   place the permissions link and remove the edit-based links - reference stylesheet asset instead of embedded - routes in config/ and fixed to work with 1.4 onward - add return link from permissions edit page - change urls to be consistent with redmine
