+++ app/views/hooks/_html_header_rwp.html.erb
@@ -1,0 +1,3 @@
+<% if controller and controller.class.name == 'WikiController' and controller.action_name == "show" %>
+    <%= stylesheet_link_tag "wiki_permissions.css", :plugin => "redmine_wiki_permissions", :media => "screen" %>
+<% end %>
